crew_ai/fixture_service/requirements.txt
# Core dependencies from the original crew AI system
crewai
crewai-tools
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
setuptools
chromadb
embedchain
google-generativeai
langchain-google-genai
Pillow
# Database dependencies
supabase>=2.0.0
psycopg2-binary>=2.9.0

# HTTP requests for API calls
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.0
